Analysis expr complexity support and simplify pass is turned on based on it.
